feat(Product service): add image file service to accept and store images anywhere

+++ EShoppingZone/Product-Service/src/main/java/com/eshopingzone/productservice/proxy/CartClient.java
@@ -1,20 +1,19 @@
 package com.eshopingzone.productservice.proxy;
 
-import java.math.BigDecimal;
+
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @FeignClient(name = "cart-service")
 public interface CartClient {
 
 	@PostMapping("/eshoppingzone/cart/addProducts")
-	void addProductsToCart(@RequestParam int userId, @RequestParam int productId, @RequestParam String productName,
-			@RequestParam String productImage, @RequestParam BigDecimal price, @RequestParam int quantity);
+	void addProductsToCart(@RequestParam int userId, @RequestParam Long productId, @RequestParam String productName,
+			@RequestParam String productImage, @RequestParam double price, @RequestParam int quantity);
 	
 	@DeleteMapping("/eshoppingzone/cart/deleteProduct")
-	void deleteProductFromCart(@RequestParam int userId, @RequestParam int productId);
+	void deleteProductFromCart(@RequestParam int userId, @RequestParam Long productId);
 }
